switched back to using single click
switched back to using single click  login without influencer explicitly specifying the username, email and password
