+++ client/src/components/Home/InfluencerHome.tsx
@@ -5,8 +5,8 @@ import Box from '@mui/material/Box'
 import ApplicationList from './ApplicationList'
 import ProfileLink from './ProfileLink'
 import MarketplaceLink from './MarketplaceLink'
-import { containerStyle } from '../../utils/sharedStyles'
 import useAuth from '../../hooks/AuthHook'
+import { containerStyle } from '../../utils/utils'
 
 export default function InfluencerHome (): JSX.Element {
   const auth = useAuth()
