--- client/src/utils/sharedStyles.ts
@@ -1,28 +1,0 @@
-export const containerStyle = {
-  contentBox: {
-    display: 'flex',
-    flexGrow: 1,
-    height: '100vh',
-    overflow: 'auto'
-  },
-  centeredBox: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center'
-  },
-  centeredPaper: {
-    p: 2,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center'
-  },
-  contentPaper: {
-    p: 2,
-    display: 'flex',
-    flexDirection: 'column'
-  },
-  contentContainer: {
-    mt: 6,
-    mb: 6
-  }
-}
